+++ ci/notes/CHANGELOG.md
@@ -4,4 +4,5 @@ We have enhanced the phlow so it's better to use it when building pipelines supp
 
 #### Features
 - upnext now has a prefix option #79 @groenborg
+- deliver now always change back to your default branch #75 @groenborg